% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amzn_frequent_terms.R
\name{amzn_frequent_terms}
\alias{amzn_frequent_terms}
\title{Most Frequent Terms by Star Rating}
\usage{
amzn_frequent_terms(data, count)
}
\arguments{
\item{data}{Data frame created by \code{GSPtext::get_reviews()} or in the same format}

\item{count}{Number of terms to return per star rating}
}
\value{
Returns a list with two items - item 1 ($data) is the data frame with top terms for each star rating; item 2 ($graph) is a ggplot bar graph faceted by star rating
}
\description{
Function to calculate the top terms for each star rating.
}
\examples{
\dontrun{
out <- amzn_frequent_terms(data, 15)
# to access data
top_terms <- out$data

# to call the graph
out$graph
}
}
