% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amzn_review_wc.R
\name{amzn_review_wc}
\alias{amzn_review_wc}
\title{Review Wordclouds}
\usage{
amzn_review_wc(
  data,
  type = c("overall", "comparison"),
  comp_type = c("rating", "sentiment")
)
}
\arguments{
\item{data}{Data frame created by \code{GSPtext::get_reviews()} or in the same format}

\item{type}{Type of word cloud, either "overall" or "comparison"}

\item{comp_type}{If "comparison," then either "ratings" (1&2 star vs 4&5 star) or "sentiment" (based on the Bing lexicon)}
}
\value{
Wordcloud
}
\description{
Tokenizing text and plotting in wordclouds. Two comparison wordclouds available ("ratings" and "sentiment")

This uses the \code{wordcloud::wordcloud()} function so we have to specify the format and save type beforehand. See examples below
}
\examples{
\dontrun{
amzn_review_wc(data, type = "comparison", comp_type = "rating"),
amzn_review_wc(data, type = "overall")

# To save a wordlcoud:
png("~/Desktop/wordcloud.png") # or any other directory/filename
amzn_review_wc(data, type = "comparison", comp_type = "rating")
dev.off()}
}
