% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_url}
\alias{build_url}
\title{Build URL}
\usage{
build_url(link)
}
\arguments{
\item{link}{A link to an Amazon product}
}
\value{
Character string - properly formatted url with pagination
}
\description{
Provide a url to an Amazon product and function creates the proper paginated review url. The url can include any number of parameters.
}
\examples{
\dontrun{
build_url(
"https://www.amazon.com/AmazonBasics-Nespresso-Storage-Capsule-Capacity/dp/B010RLCH2U/ref=xxxx")}
}
