% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reviews.R
\name{get_reviews}
\alias{get_reviews}
\title{Get Reviews}
\usage{
get_reviews()
}
\value{
Data frame with date, stars, headline, review text, and the paginated link.
}
\description{
Convenience wrapper around functions that scrape Amazon reviews. Provide a product link and this will determine the number of pages to crawl over and will return a clean data frame of reviews.
The function provides a prompt for the user to paste in the product url.
}
\examples{
\dontrun{
get_reviews()}
}
