% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_count.R
\name{page_count}
\alias{page_count}
\title{Page Count}
\usage{
page_count(url)
}
\arguments{
\item{url}{URL built with \code{build_url()} function}
}
\value{
Numeric value of total pages
}
\description{
Takes first page of Amazon product reviews and determines how many pages of reviews are available
}
\examples{
\dontrun{
page_count(url)}
}
