% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{review_summary}
\alias{review_summary}
\title{Summary Review}
\usage{
review_summary(data)
}
\arguments{
\item{data}{Data frame created by \code{GSPtext::get_reviews()} or in the same format}
}
\value{
Information is printed in console for summary purposes
}
\description{
Summarizing the reviews scraped.

Console prints the number of reviews scraped, average word count,
}
\examples{
\dontrun{
review_summary(data)}
}
