% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_term_co.R
\name{review_term_co}
\alias{review_term_co}
\title{Graph Co-Occurrences}
\usage{
review_term_co(data, star, nn)
}
\arguments{
\item{data}{Data frame created by \code{GSPtext::get_reviews()} or in the same format}

\item{star}{Star rating of reviews that should be plotted}

\item{nn}{Minimum number of co-occurrences that should be used in the plot}
}
\value{
ggraph plot
}
\description{
Count and plot the number of times pairs of words co-occur within reviews.

The function requires the user to specify which star-rating to isolate and to choose how frequently terms must co-occur.
}
\examples{
\dontrun{
review_term_co(data, star = 5, nn = 15)}
}
