% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_reviews}
\alias{sample_reviews}
\title{770 customer reviews of a winter dog hoodie}
\format{
A data frame with 770 rows and 5 variables:
\describe{
\item{date}{date of review}
\item{stars}{star rating of review, 1 to 5}
\item{headline}{review headline}
\item{text}{review text}
\item{link}{url to the page of the review}
...
}
}
\source{
\url{http://www.amazon.com}
}
\usage{
sample_reviews
}
\description{
A dataset containing reviews - date, star rating, headline, and review text
of the Jecikelon Winter Dog Hoody Sweater
}
\keyword{datasets}
