% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scraper}
\alias{scraper}
\title{Scraper of Reviews}
\usage{
scraper(link)
}
\arguments{
\item{link}{String url for Amazon product}
}
\value{
List of tibbles with date of review, star rating, headline, text, and link to review
}
\description{
Function to pull Amazon reviews. Given a standard product link, it proprerly formats a paginated url, determines the number of pages to crawl, and then returns the reviews as a list
}
\examples{
\dontrun{
scraper("url")}
}
