% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_context.R
\name{term_context}
\alias{term_context}
\title{Terms in Context}
\usage{
term_context(
  data,
  pattern,
  window = 5,
  valuetype = c("glob", "regex", "fixed")
)
}
\arguments{
\item{data}{Data frame created by \code{GSPtext::get_reviews()} or in the same format}

\item{pattern}{Keyword or phrase to search for (See examples)}

\item{window}{Number of words to display on either side of the keyword match}

\item{valuetype}{Type of pattern match. "glob" for glob-style wildcards, "regex" for regular expressions, "fixed" for exact matching}
}
\value{
dataframe with 7 variables (\code{docname or index, from, to, pre, keyword, post, pattern})
}
\description{
Shamelessly borrowing the \code{quanteda::kwic()} function (\url{https://quanteda.io/index.html}) for keyword search
}
\examples{
\dontrun{
# single token
kwic(data, pattern = "quality", window = 4, valuetype = "glob")
kwic(data, pattern = "qual", valuetype = "regex", window = 4)

# phrase matching
kwic(data, pattern = "cheap quality", window = 4)
}
}
