% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_by_star.R
\name{terms_by_star}
\alias{terms_by_star}
\title{Terms by Star}
\usage{
terms_by_star(data)
}
\arguments{
\item{data}{Data frame created by \code{GSPtext::get_reviews()} or in the same format}
}
\value{
ggplot2 graph - scatterplot with overlaid text
}
\description{
Graph with the most frequent terms used and each term is positioned by its average star rating. The dashed line represents the average star rating for the product.
}
\examples{
\dontrun{
terms_by_star(data)}
}
