% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amzn_get_reviews.R
\name{amzn_get_reviews}
\alias{amzn_get_reviews}
\title{Get Amazon Reviews}
\usage{
amzn_get_reviews(link)
}
\arguments{
\item{link}{A link to an Amazon product}
}
\value{
Data frame with date, stars, headline, review text, and the paginated link.
}
\description{
Convenience wrapper around functions that scrape Amazon reviews. Provide a product link and this will determine the number of pages to crawl over and will return a clean data frame of reviews.
}
\examples{
\dontrun{
url <- "https://www.amazon.com/Sweater-Sweaters-Yorkie-Clothes-Apparel/dp/B098L4142J/ref=sr_1_24"
data <- amzn_get_reviews(url}
}
