% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_collage}
\alias{make_collage}
\title{Make Collage of UGC Images}
\usage{
make_collage(images)
}
\arguments{
\item{images}{Array of urls; each url is a UGC image link}
}
\value{
png collage; found on ~/Desktop with name \code{ugc_collage.png}
}
\description{
Using the magick package to read UGC images and create a collage

Not every review posts a photo, but row order isn't kept by the crawler. As a result, there is no mapping
between review, rating, and image; one composite is created rather than e.g., images by star rating.
}
\examples{
\dontrun{
make_collage(out_images)}
}
